build fixes

+++ iavl/internal/compaction_options.go
@@ -2,5 +2,5 @@ package internal
 
 type PruneOptions struct {
 	RetainVersion          uint32
-	CompactionRolloverSize int
+	CompactionRolloverSize int64
 }
